requirements.txt
gradio
Pillow
requests
python-dotenv